+++ src/pages/app/apikeys.vue
@@ -8,7 +8,6 @@ import { useSupabase } from '~/services/supabase'
 import type { definitions } from '~/types/supabase'
 import TitleHead from '~/components/TitleHead.vue'
 import Spinner from '~/components/Spinner.vue'
-import { createKeys } from '~/services/apikeys'
 
 const { t } = useI18n()
 const isLoading = ref(false)
@@ -34,13 +33,12 @@ const geKeys = async (retry = true): Promise<void> => {
     .from<definitions['apikeys']>('apikeys')
     .select()
     .eq('user_id', auth?.id)
-  if (data && data.length) {
+  if (data && data.length)
     apps.value = data
-  }
-  else if (retry && auth?.id) {
-    await createKeys(auth?.id)
+
+  else if (retry && auth?.id)
     return geKeys(false)
-  }
+
   isLoading.value = false
 }
 watchEffect(async () => {