+++ src/types/supabase.ts
@@ -1222,6 +1222,8 @@ export interface paths {
           abtest?: parameters["rowFilter.plans.abtest"];
           progressive_deploy?: parameters["rowFilter.plans.progressive_deploy"];
           id?: parameters["rowFilter.plans.id"];
+          price_m_id?: parameters["rowFilter.plans.price_m_id"];
+          price_y_id?: parameters["rowFilter.plans.price_y_id"];
           /** Filtering Columns */
           select?: parameters["select"];
           /** Ordering */
@@ -1287,6 +1289,8 @@ export interface paths {
           abtest?: parameters["rowFilter.plans.abtest"];
           progressive_deploy?: parameters["rowFilter.plans.progressive_deploy"];
           id?: parameters["rowFilter.plans.id"];
+          price_m_id?: parameters["rowFilter.plans.price_m_id"];
+          price_y_id?: parameters["rowFilter.plans.price_y_id"];
         };
         header: {
           /** Preference */
@@ -1316,6 +1320,8 @@ export interface paths {
           abtest?: parameters["rowFilter.plans.abtest"];
           progressive_deploy?: parameters["rowFilter.plans.progressive_deploy"];
           id?: parameters["rowFilter.plans.id"];
+          price_m_id?: parameters["rowFilter.plans.price_m_id"];
+          price_y_id?: parameters["rowFilter.plans.price_y_id"];
         };
         body: {
           /** plans */
@@ -1566,6 +1572,7 @@ export interface paths {
           product_id?: parameters["rowFilter.stripe_info.product_id"];
           trial_at?: parameters["rowFilter.stripe_info.trial_at"];
           price_id?: parameters["rowFilter.stripe_info.price_id"];
+          is_good_plan?: parameters["rowFilter.stripe_info.is_good_plan"];
           /** Filtering Columns */
           select?: parameters["select"];
           /** Ordering */
@@ -1624,6 +1631,7 @@ export interface paths {
           product_id?: parameters["rowFilter.stripe_info.product_id"];
           trial_at?: parameters["rowFilter.stripe_info.trial_at"];
           price_id?: parameters["rowFilter.stripe_info.price_id"];
+          is_good_plan?: parameters["rowFilter.stripe_info.is_good_plan"];
         };
         header: {
           /** Preference */
@@ -1646,6 +1654,7 @@ export interface paths {
           product_id?: parameters["rowFilter.stripe_info.product_id"];
           trial_at?: parameters["rowFilter.stripe_info.trial_at"];
           price_id?: parameters["rowFilter.stripe_info.price_id"];
+          is_good_plan?: parameters["rowFilter.stripe_info.is_good_plan"];
         };
         body: {
           /** stripe_info */
@@ -1804,6 +1813,28 @@ export interface paths {
       };
     };
   };
+  "/rpc/is_in_channel": {
+    post: {
+      parameters: {
+        body: {
+          args: {
+            /** Format: uuid */
+            ownerid: string;
+            /** Format: uuid */
+            userid: string;
+          };
+        };
+        header: {
+          /** Preference */
+          Prefer?: parameters["preferParams"];
+        };
+      };
+      responses: {
+        /** OK */
+        200: unknown;
+      };
+    };
+  };
   "/rpc/get_max_channel": {
     post: {
       parameters: {
@@ -1907,6 +1938,34 @@ export interface paths {
             apikey: string;
             /** Format: character varying */
             appid: string;
+          };
+        };
+        header: {
+          /** Preference */
+          Prefer?: parameters["preferParams"];
+        };
+      };
+      responses: {
+        /** OK */
+        200: unknown;
+      };
+    };
+  };
+  "/rpc/find_fit_plan": {
+    post: {
+      parameters: {
+        body: {
+          args: {
+            /** Format: bigint */
+            apps_n: number;
+            /** Format: bigint */
+            channels_n: number;
+            /** Format: bigint */
+            shared_n: number;
+            /** Format: bigint */
+            updates_n: number;
+            /** Format: bigint */
+            versions_n: number;
           };
         };
         header: {
@@ -2115,8 +2174,30 @@ export interface paths {
       parameters: {
         body: {
           args: {
+            /** Format: character varying */
+            dateid: string;
             /** Format: uuid */
             userid: string;
+          };
+        };
+        header: {
+          /** Preference */
+          Prefer?: parameters["preferParams"];
+        };
+      };
+      responses: {
+        /** OK */
+        200: unknown;
+      };
+    };
+  };
+  "/rpc/exist_user": {
+    post: {
+      parameters: {
+        body: {
+          args: {
+            /** Format: character varying */
+            e_mail: string;
           };
         };
         header: {
@@ -2174,6 +2255,26 @@ export interface paths {
       };
     };
   };
+  "/rpc/is_allowed_action": {
+    post: {
+      parameters: {
+        body: {
+          args: {
+            /** Format: text */
+            apikey: string;
+          };
+        };
+        header: {
+          /** Preference */
+          Prefer?: parameters["preferParams"];
+        };
+      };
+      responses: {
+        /** OK */
+        200: unknown;
+      };
+    };
+  };
   "/rpc/get_dl_by_month": {
     post: {
       parameters: {
@@ -2211,7 +2312,11 @@ export interface definitions {
      * @default now()
      */
     created_at?: string;
-    /** Format: uuid */
+    /**
+     * Format: uuid
+     * @description Note:
+     * This is a Foreign Key to `users.id`.<fk table='users' column='id'/>
+     */
     user_id: string;
     /** Format: character varying */
     key: string;
@@ -2274,7 +2379,11 @@ export interface definitions {
      * This is a Foreign Key to `apps.app_id`.<fk table='apps' column='app_id'/>
      */
     app_id: string;
-    /** Format: uuid */
+    /**
+     * Format: uuid
+     * @description Note:
+     * This is a Foreign Key to `users.id`.<fk table='users' column='id'/>
+     */
     user_id: string;
     /**
      * Format: timestamp with time zone
@@ -2319,7 +2428,11 @@ export interface definitions {
     name: string;
     /** Format: character varying */
     bucket_id?: string;
-    /** Format: uuid */
+    /**
+     * Format: uuid
+     * @description Note:
+     * This is a Foreign Key to `users.id`.<fk table='users' column='id'/>
+     */
     user_id: string;
     /**
      * Format: timestamp with time zone
@@ -2348,7 +2461,11 @@ export interface definitions {
     app_id: string;
     /** Format: character varying */
     icon_url: string;
-    /** Format: uuid */
+    /**
+     * Format: uuid
+     * @description Note:
+     * This is a Foreign Key to `users.id`.<fk table='users' column='id'/>
+     */
     user_id: string;
     /** Format: character varying */
     name?: string;
@@ -2390,7 +2507,7 @@ export interface definitions {
      * @description Note:
      * This is a Foreign Key to `users.id`.<fk table='users' column='id'/>
      */
-    created_by?: string;
+    created_by: string;
     /**
      * Format: text
      * @description Note:
@@ -2670,6 +2787,10 @@ export interface definitions {
      * @default extensions.uuid_generate_v4()
      */
     id: string;
+    /** Format: character varying */
+    price_m_id: string;
+    /** Format: character varying */
+    price_y_id: string;
   };
   stats: {
     /**
@@ -2766,6 +2887,7 @@ export interface definitions {
      * Format: character varying
      * @description Note:
      * This is a Primary Key.<pk/>
+     * This is a Foreign Key to `users.customer_id`.<fk table='users' column='customer_id'/>
      */
     customer_id: string;
     /**
@@ -2781,7 +2903,8 @@ export interface definitions {
       | "canceled";
     /**
      * Format: character varying
-     * @default free
+     * @description Note:
+     * This is a Foreign Key to `plans.stripe_id`.<fk table='plans' column='stripe_id'/>
      */
     product_id: string;
     /**
@@ -2791,6 +2914,11 @@ export interface definitions {
     trial_at: string;
     /** Format: character varying */
     price_id?: string;
+    /**
+     * Format: boolean
+     * @default true
+     */
+    is_good_plan?: boolean;
   };
   users: {
     /**
@@ -2834,7 +2962,11 @@ export interface definitions {
      * @default false
      */
     legalAccepted: boolean;
-    /** Format: character varying */
+    /**
+     * Format: character varying
+     * @description Note:
+     * This is a Foreign Key to `stripe_info.customer_id`.<fk table='stripe_info' column='customer_id'/>
+     */
     customer_id?: string;
   };
 }
@@ -3104,6 +3236,10 @@ export interface parameters {
   "rowFilter.plans.progressive_deploy": string;
   /** Format: uuid */
   "rowFilter.plans.id": string;
+  /** Format: character varying */
+  "rowFilter.plans.price_m_id": string;
+  /** Format: character varying */
+  "rowFilter.plans.price_y_id": string;
   /** @description stats */
   "body.stats": definitions["stats"];
   /** Format: bigint */
@@ -3162,6 +3298,8 @@ export interface parameters {
   "rowFilter.stripe_info.trial_at": string;
   /** Format: character varying */
   "rowFilter.stripe_info.price_id": string;
+  /** Format: boolean */
+  "rowFilter.stripe_info.is_good_plan": string;
   /** @description users */
   "body.users": definitions["users"];
   /** Format: timestamp with time zone */