+++ supabase/functions/demo/index.ts
@@ -1,4 +1,4 @@
-import { serve } from 'https://deno.land/std@0.139.0/http/server.ts'
+import { serve } from 'https://deno.land/std@0.143.0/http/server.ts'
 import { supabaseAdmin } from '../_utils/supabase.ts'
 import type { definitions } from '../_utils/types_supabase.ts'
 import { checkKey, sendRes } from '../_utils/utils.ts'
@@ -10,7 +10,7 @@ interface dataDemo {
   iconType: string
 }
 
-serve(async(event: Request) => {
+serve(async (event: Request) => {
   const supabase = supabaseAdmin
   const apikey_string = event.headers.get('apikey')
   if (!apikey_string)