+++ supabase/functions/stripe_portal/index.ts
@@ -1,4 +1,4 @@
-import { serve } from 'https://deno.land/std@0.139.0/http/server.ts'
+import { serve } from 'https://deno.land/std@0.143.0/http/server.ts'
 import { createPortal } from '../_utils/stripe.ts'
 import { supabaseAdmin } from '../_utils/supabase.ts'
 import type { definitions } from '../_utils/types_supabase.ts'
@@ -8,7 +8,7 @@ interface PortalData {
   callbackUrl: string
 }
 
-serve(async(event: Request) => {
+serve(async (event: Request) => {
   console.log('method', event.method)
   if (event.method === 'OPTIONS')
     return sendOptionsRes()
@@ -17,15 +17,16 @@ serve(async(event: Request) => {
   if (!authorization)
     return sendRes({ status: 'Cannot find authorization' }, 400)
   try {
-    let body: PortalData = {callbackUrl: `${Deno.env.get('WEBAPP_URL')}/app/usage`}
+    let body: PortalData = { callbackUrl: `${Deno.env.get('WEBAPP_URL')}/app/usage` }
     try {
       body = (await event.json()) as PortalData
     // deno-lint-ignore no-empty
-    } catch {}
+    }
+    catch {}
     const { user: auth, error } = await supabase.auth.api.getUser(
       authorization?.split('Bearer ')[1],
     )
-    // eslint-disable-next-line no-console
+
     console.log('auth', auth)
     if (error || !auth)
       return sendRes({ status: 'not authorize' }, 400)
@@ -39,7 +40,7 @@ serve(async(event: Request) => {
       return sendRes({ status: 'not authorize' }, 400)
     if (!user.customer_id)
       return sendRes({ status: 'no customer' }, 400)
-    // eslint-disable-next-line no-console
+
     console.log('user', user)
     const link = await createPortal(Deno.env.get('STRIPE_SECRET_KEY') || '', user.customer_id, body.callbackUrl)
     return sendRes({ url: link.url })